do not call emergency many time

+++ src/VESPA/VESPA_module.py
@@ -868,16 +868,17 @@ class Drone:
             self.change_state_to(Irremovable)
 
     def return_home(self, vehicle):
-        if self.id==1: # Sink
-            #time.sleep(10) 
-            vehicle.mode = VehicleMode ("LAND")
-        else:
-            #time.sleep(5 * (self.id)) # Wait time proportional to the id so not all back to home at the same time 
-            vehicle.mode = VehicleMode ("RTL")
+        if vehicle is not None:
+            if self.id==1: # Sink
+                #time.sleep(10) 
+                vehicle.mode = VehicleMode ("LAND")
+            else:
+                #time.sleep(5 * (self.id)) # Wait time proportional to the id so not all back to home at the same time 
+                vehicle.mode = VehicleMode ("RTL")
         
 
     def interrupt(self, vehicle):
-            if vehicle is not None:
+            if not self.Emergency_stop.is_set():
                 emergency_msg= self.build_emergency_message()
                 send_msg(emergency_msg)
                 print("retuen home")
@@ -913,19 +914,18 @@ class Drone:
                 sys.exit(1)  # Exit the program with a non-zero status
  
     def emergency_stop(self):
-        print("Emergency stop detected. Exiting function.")
-        # brodcast it again
-        emergency_msg= self.build_emergency_message()
-        send_msg(emergency_msg)
-        print("Retuen home")
-        threads = threading.enumerate()
-        print(f"emergency_stop Number of active threads: {len(threads)}")
-        for thread in threads:
-            print(f"Thread name: {thread.name or 'Unnamed'}, "
-                f"Thread ID: {thread.ident}, "
-                f"Daemon: {thread.daemon}, "
-                f"Alive: {thread.is_alive()}")
-
-
-        os.kill(os.getpid(), signal.SIGINT) # That will call interrupt which use vehicle object to return home
-        #self.interrupt(vehicle)
+        if not self.Emergency_stop.is_set():
+            print("Emergency stop detected. Exiting function.")
+            # brodcast it again
+            emergency_msg= self.build_emergency_message()
+            send_msg(emergency_msg)
+            print("Retuen home")
+            threads = threading.enumerate()
+            print(f"emergency_stop Number of active threads: {len(threads)}")
+            for thread in threads:
+                print(f"Thread name: {thread.name or 'Unnamed'}, "
+                    f"Thread ID: {thread.ident}, "
+                    f"Daemon: {thread.daemon}, "
+                    f"Alive: {thread.is_alive()}")
+            os.kill(os.getpid(), signal.SIGINT) # That will call interrupt which use vehicle object to return home
+            #self.interrupt(vehicle)
